Added arguments to calcBaselinL to be able to use with AIRR or Change-O
When I changed defaults to AIRR format, unit testing for calcBaseline would fail due to
not finding junction_length and clone_id in the Change-O formatted example data. The problem was
 calcBaseline calls calcBaselineOneClone, then calcBaselineL (which didn't take as arguments junction
length or clone id columns), then obervedMutations (which is some cases, needed to know the right
names for the junction length and clone id columns). So I added `cloneColumn` and `juncLengthColumn` to
`calcBaselineL`. Still to do: standardize the use of `juncLengthColumn` or `juncLenCol`, and refactor
`calcBaseline`+`calcBaselineL`.