Changes to avoid warnings in testthat
I moved to the code to ensure older version of sample() from
being at the to of the file, to being specified in set.seed to
avoid a warning that was something like 'you are setting
rounding outside of test_that'. I also added expect_warning,
because set.seed uses a warning to inform "non-uniform
'Rounding' sampler used".